Move Dummy benchmark into separate project
This gets rid of the org.renaissance.core package and makes Dummy
serve its primary purpose, i.e., to provide a simple way to test the
packaging and class-loading mechanism.

This also removes the unfortunate use of the Dummy benchmark by the
Policy class. A local dummy instance has been put in place temporarily
before the Policy class is cleaned up.

Finally, this cleans up a little the handling of classes in
listBenchmarks() in build.sbt, where we exclude all classes that
come from org.renaissance, org.renaissance.harness and
org.renaissance.util packages, because these cannot contain any
benchmarks.
