+++ benchmarks/dummy/src/main/java/org/renaissance/dummy/Dummy.java
@@ -1,4 +1,4 @@
-package org.renaissance.core;
+package org.renaissance.dummy;
 
 import org.renaissance.Config;
 import org.renaissance.License;
@@ -6,7 +6,7 @@ import org.renaissance.RenaissanceBenchmark;
 
 import static org.renaissance.License.MIT;
 
-public class Dummy extends RenaissanceBenchmark {
+public final class Dummy extends RenaissanceBenchmark {
   @Override
   public String description() {
     return "A dummy benchmark, which does no work. It is used only to test the harness.";
